backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
yt-dlp==2023.11.16
openai-whisper==20231117
torch==2.1.1
transformers==4.35.2
TTS==0.22.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
argostranslate==1.9.1 